Improve negated reconciliation logic
