+++ src/Psalm/Internal/Type/NegatedAssertionReconciler.php
@@ -7,10 +7,6 @@ use Psalm\Internal\Analyzer\StatementsAnalyzer;
 use Psalm\Internal\Analyzer\TraitAnalyzer;
 use Psalm\Internal\Type\Comparator\AtomicTypeComparator;
 use Psalm\Internal\Type\Comparator\UnionTypeComparator;
-use Psalm\Issue\DocblockTypeContradiction;
-use Psalm\Issue\RedundantPropertyInitializationCheck;
-use Psalm\Issue\TypeDoesNotContainType;
-use Psalm\IssueBuffer;
 use Psalm\Type;
 use Psalm\Type\Atomic\TArray;
 use Psalm\Type\Atomic\TEmptyMixed;
@@ -20,7 +16,6 @@ use Psalm\Type\Atomic\TFloat;
 use Psalm\Type\Atomic\TInt;
 use Psalm\Type\Atomic\TIterable;
 use Psalm\Type\Atomic\TNamedObject;
-use Psalm\Type\Atomic\TNonEmptyMixed;
 use Psalm\Type\Atomic\TNonEmptyString;
 use Psalm\Type\Atomic\TString;
 use Psalm\Type\Atomic\TTrue;
@@ -82,113 +77,6 @@ class NegatedAssertionReconciler extends Reconciler
             return $existing_var_type;
         }
 
-        if (!$is_equality) {
-            if ($assertion === 'isset') {
-                if ($existing_var_type->possibly_undefined) {
-                    return Type::getEmpty();
-                }
-
-                if (!$existing_var_type->isNullable()
-                    && $key
-                    && strpos($key, '[') === false
-                ) {
-                    foreach ($existing_var_type->getAtomicTypes() as $atomic) {
-                        if (!$existing_var_type->hasMixed()
-                            || $atomic instanceof TNonEmptyMixed
-                        ) {
-                            $failed_reconciliation = Reconciler::RECONCILIATION_EMPTY;
-
-                            if ($code_location) {
-                                if ($existing_var_type->from_static_property) {
-                                    IssueBuffer::maybeAdd(
-                                        new RedundantPropertyInitializationCheck(
-                                            'Static property ' . $key . ' with type '
-                                                . $existing_var_type
-                                                . ' has unexpected isset check — should it be nullable?',
-                                            $code_location
-                                        ),
-                                        $suppressed_issues
-                                    );
-                                } elseif ($existing_var_type->from_property) {
-                                    IssueBuffer::maybeAdd(
-                                        new RedundantPropertyInitializationCheck(
-                                            'Property ' . $key . ' with type '
-                                                . $existing_var_type . ' should already be set in the constructor',
-                                            $code_location
-                                        ),
-                                        $suppressed_issues
-                                    );
-                                } elseif ($existing_var_type->from_docblock) {
-                                    IssueBuffer::maybeAdd(
-                                        new DocblockTypeContradiction(
-                                            'Cannot resolve types for ' . $key . ' with docblock-defined type '
-                                                . $existing_var_type . ' and !isset assertion',
-                                            $code_location,
-                                            null
-                                        ),
-                                        $suppressed_issues
-                                    );
-                                } else {
-                                    IssueBuffer::maybeAdd(
-                                        new TypeDoesNotContainType(
-                                            'Cannot resolve types for ' . $key . ' with type '
-                                                . $existing_var_type . ' and !isset assertion',
-                                            $code_location,
-                                            null
-                                        ),
-                                        $suppressed_issues
-                                    );
-                                }
-                            }
-
-                            return $existing_var_type->from_docblock
-                                ? Type::getNull()
-                                : Type::getEmpty();
-                        }
-                    }
-                }
-
-                return Type::getNull();
-            }
-
-            if ($assertion === 'array-key-exists') {
-                return Type::getEmpty();
-            }
-
-            if (strpos($assertion, 'in-array-') === 0) {
-                $assertion = substr($assertion, 9);
-                $new_var_type = Type::parseString($assertion);
-
-                $intersection = Type::intersectUnionTypes(
-                    $new_var_type,
-                    $existing_var_type,
-                    $statements_analyzer->getCodebase()
-                );
-
-                if ($intersection === null) {
-                    if ($key && $code_location) {
-                        self::triggerIssueForImpossible(
-                            $existing_var_type,
-                            $existing_var_type->getId(),
-                            $key,
-                            '!' . $assertion,
-                            true,
-                            $negated,
-                            $code_location,
-                            $suppressed_issues
-                        );
-                    }
-
-                    $failed_reconciliation = Reconciler::RECONCILIATION_EMPTY;
-                }
-
-                return $existing_var_type;
-            }
-
-            if (strpos($assertion, 'has-array-key-') === 0) {
-                return $existing_var_type;
-            }
-        }
 
         $existing_var_atomic_types = $existing_var_type->getAtomicTypes();
 
@@ -200,6 +88,7 @@ class NegatedAssertionReconciler extends Reconciler
             $existing_var_type->addType(new TFalse);
         } else {
             $simple_negated_type = SimpleNegatedAssertionReconciler::reconcile(
+                $statements_analyzer->getCodebase(),
                 $assertion,
                 $existing_var_type,
                 $key,
